Rename package to 'dev.sjaramillo.pedometer'.
