+++ app/src/main/java/dev/sjaramillo/pedometer/BootReceiver.java
@@ -14,26 +14,45 @@
  * limitations under the License.
  */
 
-package de.j4velin.pedometer;
+package dev.sjaramillo.pedometer;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Build;
 
-import de.j4velin.pedometer.util.API26Wrapper;
-import de.j4velin.pedometer.util.Logger;
+import dev.sjaramillo.pedometer.util.API26Wrapper;
+import dev.sjaramillo.pedometer.util.Logger;
 
-public class AppUpdatedReceiver extends BroadcastReceiver {
+public class BootReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(final Context context, final Intent intent) {
-        if (BuildConfig.DEBUG) Logger.log("app updated");
+        if (BuildConfig.DEBUG) Logger.log("booted");
+
+        SharedPreferences prefs = context.getSharedPreferences("pedometer", Context.MODE_PRIVATE);
+
+        Database db = Database.getInstance(context);
+
+        if (!prefs.getBoolean("correctShutdown", false)) {
+            if (BuildConfig.DEBUG) Logger.log("Incorrect shutdown");
+            // can we at least recover some steps?
+            int steps = Math.max(0, db.getCurrentSteps());
+            if (BuildConfig.DEBUG) Logger.log("Trying to recover " + steps + " steps");
+            db.addToLastEntry(steps);
+        }
+        // last entry might still have a negative step value, so remove that
+        // row if that's the case
+        db.removeNegativeEntries();
+        db.saveCurrentSteps(0);
+        db.close();
+        prefs.edit().remove("correctShutdown").apply();
+        
         if (Build.VERSION.SDK_INT >= 26) {
             API26Wrapper.startForegroundService(context, new Intent(context, SensorListener.class));
         } else {
             context.startService(new Intent(context, SensorListener.class));
         }
     }
-
 }
