+++ app/src/main/java/dev/sjaramillo/pedometer/Database.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.j4velin.pedometer;
+package dev.sjaramillo.pedometer;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -28,8 +28,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import de.j4velin.pedometer.util.Logger;
-import de.j4velin.pedometer.util.Util;
+import dev.sjaramillo.pedometer.util.Logger;
+import dev.sjaramillo.pedometer.util.Util;
 
 public class Database extends SQLiteOpenHelper {
