+++ app/src/main/res/layout-land/fragment_overview.xml
@@ -3,7 +3,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:padding="5dp"
-                tools:context=".Activity_Main">
+                tools:context=".ui.Activity_Main">
 
     <org.eazegraph.lib.charts.PieChart
         xmlns:eaze="http://schemas.android.com/apk/res-auto"
